+++ plugins/catalog/src/components/CatalogPage/DefaultCatalogPage.tsx
@@ -28,6 +28,7 @@ import {
   CatalogFilterLayout,
   EntityLifecyclePicker,
   EntityListProvider,
+  EntityProcessingStatusPicker,
   EntityOwnerPicker,
   EntityTagPicker,
   EntityTypePicker,
@@ -84,6 +85,7 @@ export function DefaultCatalogPage(props: DefaultCatalogPageProps) {
               <EntityOwnerPicker />
               <EntityLifecyclePicker />
               <EntityTagPicker />
+              <EntityProcessingStatusPicker />
             </CatalogFilterLayout.Filters>
             <CatalogFilterLayout.Content>
               <CatalogTable