+++ plugins/kubernetes-backend/src/service/KubernetesBuilder.test.ts
@@ -34,11 +34,13 @@ import {
 import { KubernetesBuilder } from './KubernetesBuilder';
 import { KubernetesFanOutHandler } from './KubernetesFanOutHandler';
 import { PodStatus } from '@kubernetes/client-node';
+import { CatalogApi } from '@backstage/catalog-client';
 
 describe('KubernetesBuilder', () => {
   let app: express.Express;
   let kubernetesFanOutHandler: jest.Mocked<KubernetesFanOutHandler>;
   let config: Config;
+  let catalogApi: CatalogApi;
 
   beforeAll(async () => {
     const logger = getVoidLogger();
@@ -80,6 +82,7 @@ describe('KubernetesBuilder', () => {
       config,
       logger,
       discovery,
+      catalogApi,
     })
       .setObjectsProvider(kubernetesFanOutHandler)
       .setClusterSupplier(clusterSupplier)
@@ -256,6 +259,7 @@ describe('KubernetesBuilder', () => {
         logger,
         config,
         discovery,
+        catalogApi,
       })
         .setClusterSupplier(clusterSupplier)
         .setServiceLocator(serviceLocator)