+++ plugins/search-backend-module-elasticsearch/src/engines/index.ts
@@ -15,13 +15,25 @@
  */
 
 export { ElasticSearchSearchEngine } from './ElasticSearchSearchEngine';
+export { isOpenSearchCompatible } from './ElasticSearchClientOptions';
 export type {
+  BaseElasticSearchClientOptions,
   ElasticSearchAgentOptions,
   ElasticSearchConnectionConstructor,
+  ElasticSearchElasticSearchClientOptions,
   ElasticSearchTransportConstructor,
   ElasticSearchNodeOptions,
   ElasticSearchAuth,
+  OpenSearchAuth,
+  OpenSearchConnectionConstructor,
+  OpenSearchElasticSearchClientOptions,
+  OpenSearchNodeOptions,
 } from './ElasticSearchClientOptions';
+export type {
+  ElasticSearchAliasAction,
+  ElasticSearchClientWrapper,
+  ElasticSearchIndexAction,
+} from './ElasticSearchClientWrapper';
 export type {
   ElasticSearchConcreteQuery,
   ElasticSearchClientOptions,
@@ -30,10 +42,12 @@ export type {
   ElasticSearchQueryTranslator,
   ElasticSearchQueryTranslatorOptions,
   ElasticSearchOptions,
-  ElasticSearchCustomIndexTemplate,
-  ElasticSearchCustomIndexTemplateBody,
 } from './ElasticSearchSearchEngine';
 export type {
   ElasticSearchSearchEngineIndexer,
   ElasticSearchSearchEngineIndexerOptions,
 } from './ElasticSearchSearchEngineIndexer';
+export type {
+  ElasticSearchCustomIndexTemplate,
+  ElasticSearchCustomIndexTemplateBody,
+} from './types';