+++ plugins/permission-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-permission-backend
 
+## 0.5.9-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.1
+  - @backstage/errors@1.1.0-next.0
+  - @backstage/plugin-auth-node@0.2.3-next.1
+  - @backstage/plugin-permission-common@0.6.3-next.0
+  - @backstage/plugin-permission-node@0.6.3-next.1
+
 ## 0.5.9-next.0
 
 ### Patch Changes